requirements.txt
pyodbc==4.0.39
psycopg2-binary==2.9.7
pandas==2.0.3
sqlparse==0.4.4
sqlalchemy==2.0.19
pydantic==2.1.1
PyYAML==6.0.1
numpy==1.24.3